Fixes two -Wpointer-to-int-cast warnings, from gcc 10's -Wall, on 64-bit builds
See
    eline.c:282:      if (window_vector[i] == line) window_vector[i] = (linestr *)(+1);

and comments above it:

    ... We can't set the value to NULL, as that implies the
    screen line is blank. Use (+1) and assume that will never be a valid line
    address... */

and in ekey.c:

    ... the window_vector pointer is set to +1 to indicate "something unknown
    is on the screen". */

edisplay.c:74:27: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   74 |       if (line != NULL && (int)line > 1) eoflastline = ((line->flags & lf_eof) != 0);
      |                           ^
ekey.c:399:23: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  399 |   if (line != NULL && (int)line > 1) line->flags |= lf_shn;
      |                       ^

Re-work the fixes for two -Wpointer-to-int-cast warnings, using ne-3.x

intptr_t is defined in <stdint.h>

+++ src/edisplay.c
@@ -71,7 +71,7 @@ switch (option)
       {
       line = window_vector[i];
       if (line == main_current) row = i;
-      if (line != NULL && (int)line > 1) eoflastline = ((line->flags & lf_eof) != 0);
+      if (line != NULL && (intptr_t)line > 1) eoflastline = ((line->flags & lf_eof) != 0);
       }
 
     if (row < 0) return;                /* current line off screen */
